+++ .gitignore
@@ -1,2 +1,41 @@
 .DS_Store
-/Manifest.toml
+
+*.jl.cov
+*.jl.*.cov
+*.jl.mem
+
+docs/build/
+docs/site/
+docs/src/examples/*.md
+
+*.ipynb_checkpoints
+**/*.ipynb_checkpoints
+**/**/*.ipynb_checkpoints
+
+_*.dat
+*.swp
+__pycache__/
+
+Manifest.toml
+
+*.aux
+*.log
+*.out
+*.synctex.gz
+*.fls
+*.fdb_latexmk
+*.blg
+*.bbl
+*.pyc
+*.end
+*~
+_*.tex
+*.pdf
+_*.bib
+_temp/
+.DS_store
+.ipynb_checkpoints/
+docs/_minted-quantum/
+*.toc
+*.mtc
+*.mtc0
