+++ Project.toml
@@ -3,10 +3,13 @@ uuid = "e06ad7bc-5795-54bd-b84a-0284c0717521"
 
 [deps]
 BitBasis = "50ba71b6-fa0f-514d-ae9a-0916efc90dcf"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LuxurySparse = "d05aeea4-b7d4-55ac-b691-9e7fabb07ba2"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
 Yao = "5872b779-8223-5990-8dd0-5abbb0748c8c"
 YaoBase = "a8f54c17-34bc-5a9d-b050-f522fe3f755f"
+YaoBlocks = "418bc28f-b43b-5e0b-a6e7-61bbc1a2c1df"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [extras]
