+++ README.md
@@ -1,1 +1,13 @@
 # YaoFlux
+
+Automatic differentiable quantum circuit.
+
+## Install
+```julia
+pkg> add git@github.com:GiggleLiu/YaoFlux.jl.git#vqe
+```
+
+## Run an example
+```julia
+julia examples/vqe.jl
+```
