+++ examples/vqe.jl
@@ -1,15 +1,20 @@
 # generate a Heisenberg Model Hamiltonian
-using QuAlgorithmZoo: heisenberg
+using Yao
+using YaoFlux
+using LinearAlgebra
+using QuAlgorithmZoo: heisenberg, random_diff_circuit, pair_ring
 
-h = mat(heisenberg(2))
-v0 = statevec(zero_state(2))
+nbit = 6
+h = mat(heisenberg(nbit))
+v0 = statevec(zero_state(nbit))
 function energy(circuit)
     v = mat(circuit) * v0
     (v'* h * v)[] |> real
 end
 
 # a circuit as ansatz
-circuit = chain(2, [put(2, 1=>H), put(2, 2=>H), put(2, 2=>Rz(0.0)), put(2, 2=>Rx(0.0)), control(2, 1, 2=>Z), put(2, 2=>Rx(0.0)), put(2, 2=>Rz(0.0))])
+#circuit = chain(nbit, [put(2, 1=>H), put(2, 2=>H), put(2, 2=>Rz(0.0)), put(2, 2=>Rx(0.0)), control(2, 1, 2=>Z), put(2, 2=>Rx(0.0)), put(2, 2=>Rz(0.0))])
+circuit = random_diff_circuit(nbit, 2, pair_ring(nbit))
 
 # obtain the energy
 energy(circuit)
@@ -32,4 +37,7 @@ end
 
 using Random
 Random.seed!(5)
-train!(ec4, circuit, ADAM(0.1); maxiter=100)
+EG = eigvals(Matrix(h))[1]
+println("Exact Energy is $EG")
+train!(energy, circuit, ADAM(0.1); maxiter=200)
+
